// src/harnesses/codex/protocol/generated/v2/CommandAction.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type {AbsolutePathBuf} from '../AbsolutePathBuf';

export type CommandAction =
	| {type: 'read'; command: string; name: string; path: AbsolutePathBuf}
	| {type: 'listFiles'; command: string; path: string | null}
	| {type: 'search'; command: string; query: string | null; path: string | null}
	| {type: 'unknown'; command: string};